Lab Module Write-Ups
Final documents for methods and results in Zemax introduction modules. Topics include finding the RMS spot size for a singlet lens, Design of an SLR camera lens, Young's Double Slit simulation with a comparison to theory,  design of an optical spectrometer, design of a polygon scanner, optical cryostat, and fiber coupling.

Also included is the N Slit irradiance code which was used to obtain theoretical N-slit interference patterns.